package changed to io.github.thachillera (#17)
