+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.robinkuiper.cardsscorekeeper"
+        applicationId "io.github.thachillera.cardsscorekeeper"
         minSdkVersion 25
         targetSdkVersion 28
         versionCode 1
