+++ app/src/main/res/layout/activity_gameselect.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".interfaces.GameSelectActivity">
+    tools:context="io.github.thachillera.cardsscorekeeper.interfaces.GameSelectActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
