+++ app/src/main/res/layout/activity_player_select.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".interfaces.player.PlayerSelectActivity">
+    tools:context="io.github.thachillera.cardsscorekeeper.interfaces.player.PlayerSelectActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
