+++ app/src/main/res/layout/content_gameselect.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".interfaces.GameSelectActivity"
+    tools:context="io.github.thachillera.cardsscorekeeper.interfaces.GameSelectActivity"
     tools:showIn="@layout/activity_gameselect">
 
     <Button
