+++ app/src/main/res/layout/content_player_select.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".interfaces.player.PlayerSelectActivity"
+    tools:context="io.github.thachillera.cardsscorekeeper.interfaces.player.PlayerSelectActivity"
     tools:showIn="@layout/activity_player_select">
 
     <android.support.v7.widget.RecyclerView
